pyproject.toml
[tool.poetry]
name = "ydl-wrapper"
version = "0.1.0"
description = "A wrapper around youtube-dl for easy configuration"
authors = ["Jan Willhaus <mail@janwillhaus.de>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
youtube-dl = "^2020.3.24"
logzero = "^1.5.0"
pyyaml = "^5.3.1"
colorama = "^0.4.3"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
pytest = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
